My wonderful conway emulation was broken. It was sadly uncaredfor. I have been wanting to come back and see what the issue was. So I built a test to reproduce the issue and created a "fix" for it.
Originally it seems that I was initializing LifeController by starting
with an empty call to setInterval, perhaps newere varions of js disallowed
the calling of this method without arguments. As a quick fix, I just removed
the initial call to this method upon constructin LifeController. As it turns out
I dont have a great plan for how I really want this page to setup or start.

One of the next steps I'd like to complete is making the main screen larger,
and have a better idea for how I want the page to start. I think I'd like it to
not have a starting piece set on the page. I think I'd like to have the user
definitely have to click on one of the options.